# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Futsal Reservation API
  description: |
    API complète pour l'application de réservation de matchs de futsal.
    
    Cette API permet de gérer :
    - L'authentification des utilisateurs
    - La gestion des matchs (création, inscription, désinscription)
    - Les notifications et annonces
    - L'administration des utilisateurs et statistiques
    
    ## Authentification
    L'API utilise Better-auth avec des sessions sécurisées. Les cookies sont automatiquement gérés par le navigateur.
    
    ## Rate Limiting
    - Login : 5 tentatives/minute par IP
    - Register : 3 tentatives/minute par IP
    - API générale : Protection par utilisateur authentifié
    
    ## Codes d'erreur
    - 400: Requête invalide (validation échouée)
    - 401: Non authentifié
    - 403: Accès refusé (permissions insuffisantes)
    - 404: Ressource non trouvée
    - 429: Limite de débit dépassée
    - 500: Erreur serveur interne
    
  version: 1.0.0
  contact:
    name: Futsal App Support
    email: support@futsal.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Serveur de développement
  - url: https://api.futsal.app
    description: Serveur de production
...
paths:
...
  /api/matches/{id}:
    get:
      tags:
        - Matches
      summary: Détails d'un match
      description: Récupère les détails complets d'un match avec la liste des joueurs
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID du match
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Détails du match
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MatchWithPlayers'
        '404':
          description: Match non trouvé

    put:
      tags:
        - Matches
      summary: Modifier un match (Admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMatchRequest'
      responses:
        '200':
          description: Match modifié
        '403':
          description: Accès refusé
        '404':
          description: Match non trouvé

    delete:
      tags:
        - Matches
      summary: Supprimer un match (Admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Match supprimé
        '403':
          description: Accès refusé
        '404':
          description: Match non trouvé
  /api/matches/{id}/join:
    post:
      tags:
        - Matches
      summary: S'inscrire à un match
      description: Inscrit l'utilisateur connecté au match spécifié
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Inscription réussie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          status:
                            type: string
                            enum: [confirmed, waiting]
                          position:
                            type: integer
        '400':
          description: Impossible de s'inscrire (déjà inscrit, match complet, etc.)
        '404':
          description: Match non trouvé
  /api/matches/{id}/leave:
    delete:
      tags:
        - Matches
      summary: Se désinscrire d'un match
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Désinscription réussie
        '400':
          description: Impossible de se désinscrire
        '404':
          description: Match non trouvé

  # User endpoints
  /api/user/profile:
    get:
      tags:
        - Users
      summary: Profil utilisateur
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'

    put:
      tags:
        - Users
      summary: Modifier le profil
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pseudo:
                  type: string
                  pattern: "^[a-zA-Z0-9._-]{3,20}$"
                avatar:
                  type: string
                  format: uri
      responses:
        '200':
          description: Profil mis à jour
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        pseudo:
          type: string
        avatar:
          type: string
          format: uri
        role:
          type: string
          enum: [user, root]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Match:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date-time
        maxPlayers:
          type: integer
          minimum: 2
          maximum: 24
          default: 12
        status:
          type: string
          enum: [open, full, cancelled, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MatchPlayer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        matchId:
          type: string
          format: uuid
        status:
          type: string
          enum: [confirmed, waiting]
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
    MatchWithPlayers:
      allOf:
        - $ref: '#/components/schemas/Match'
        - type: object
          properties:
            players:
              type: array
              items:
                $ref: '#/components/schemas/MatchPlayer'
...
    UpdateMatchRequest:
      type: object
      properties:
        date:
          type: string
          format: date-time
        maxPlayers:
          type: integer
          minimum: 2
          maximum: 24
        status:
          type: string
          enum: [open, full, cancelled, completed]
...
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            matchesPlayed:
              type: integer
            participationRate:
              type: number
              format: float
            badges:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  earned:
                    type: boolean
                  progress:
                    type: number
...